Refactor EntryPointManager into a separate class
