+++ editor/src/windowManagement/contentWindows/ContentWindowEntityEditor.js
@@ -1,6 +1,6 @@
 import {ContentWindow} from "./ContentWindow.js";
 import {ContentWindowOutliner} from "./ContentWindowOutliner.js";
-import {ContentWindowBuildView} from "./ContentWindowBuildView.js";
+import {ContentWindowBuildView} from "./ContentWindowBuildView/ContentWindowBuildView.js";
 import {Button} from "../../ui/Button.js";
 import {CameraComponent, ClusteredLightsConfig, Entity, GizmoManager, Mat4, OrbitControls, TranslationGizmo, Vec3} from "../../../../src/mod.js";
 import {ProjectAssetTypeEntity} from "../../assets/projectAssetType/ProjectAssetTypeEntity.js";
